relation: proof-linking: Add `LinkableCircuit` trait and impls
This trait removes the reference to proof linking in the `Circuit`
trait and allows us to implement link group layout and gate placement
generically. This will allow shared arithmetization functionality
between the multiprover and singleprover implementations.
