+++ relation/src/traits.rs
@@ -17,9 +17,7 @@ use crate::{
         EqualityGate, Gate, LinCombGate, LogicOrGate, LogicOrOutputGate, MulAddGate,
         MultiplicationGate, MuxGate, SubtractionGate,
     },
-    next_multiple,
-    proof_linking::GroupLayout,
-    BoolVar, SortedLookupVecAndPolys, Variable,
+    next_multiple, BoolVar, SortedLookupVecAndPolys, Variable,
 };
 
 /// A shorthand for converting a value to a constant in the circuit
@@ -37,26 +35,10 @@ macro_rules! felts {
     };
 }
 
-/// Represents the parameterization of a proof-linking group in the circuit
-///
-/// See `Circuit::create_link_group` for more details on proof-linking
-#[derive(Clone, Debug)]
-pub struct LinkGroup {
-    /// The id of the group
-    pub id: String,
-}
-
 /// An interface to add gates to a circuit that generalizes across wiring
 /// implementations
 ///
 /// Effectively abstracts the variable-centric gate interface
-///
-/// The generic `W` corresponds to the witness type assigned to wires, this
-/// allows both single and multiprover implementations to make use of the
-/// underlying default impls
-///
-/// The generic `C` corresponds to the constant type used to represent public
-/// values in the circuit
 ///
 /// The generic `F` corresponds to the field that the circuit is over, this
 /// type is stored in gates and used to construct selectors
@@ -138,48 +120,19 @@ pub trait Circuit<F: Field> {
 
     /// Add a constant variable to the circuit; return the index of the
     /// variable.
-    fn create_constant_variable(&mut self, val: F) -> Result<Variable, CircuitError> {
-        self.create_constant_variable_with_link_groups(val, &[])
-    }
-
-    /// Add a constant variable to the circuit; with an optional set of link
-    /// groups to add it to
-    fn create_constant_variable_with_link_groups(
-        &mut self,
-        val: F,
-        link_groups: &[LinkGroup],
-    ) -> Result<Variable, CircuitError>;
+    fn create_constant_variable(&mut self, val: Self::Constant) -> Result<Variable, CircuitError>;
 
     /// Add a variable to the circuit; return the index of the variable.
-    fn create_variable(&mut self, val: Self::Wire) -> Result<Variable, CircuitError> {
-        self.create_variable_with_link_groups(val, &[])
-    }
-
-    /// Add a variable to the witness; with an optional set of link groups to
-    /// add it to
-    fn create_variable_with_link_groups(
-        &mut self,
-        val: Self::Wire,
-        link_groups: &[LinkGroup],
-    ) -> Result<Variable, CircuitError>;
+    fn create_variable(&mut self, val: Self::Wire) -> Result<Variable, CircuitError>;
 
     /// Add a bool variable to the circuit; return the index of the variable.
     fn create_boolean_variable<T: Into<Self::Wire>>(
         &mut self,
         val: T,
     ) -> Result<BoolVar, CircuitError> {
-        self.create_boolean_variable_with_link_groups(val, &[])
-    }
-
-    /// Add a constant bool variable to the circuit; with an optional set of
-    /// link groups to add it to
-    fn create_boolean_variable_with_link_groups<T: Into<Self::Wire>>(
-        &mut self,
-        val: T,
-        link_groups: &[LinkGroup],
-    ) -> Result<BoolVar, CircuitError> {
-        let var = self.create_variable_with_link_groups(val.into(), link_groups)?;
+        let var = self.create_variable(val.into())?;
         self.enforce_bool(var)?;
+
         Ok(BoolVar(var))
     }
 
@@ -202,17 +155,6 @@ pub trait Circuit<F: Field> {
 
     /// Set a variable to a public variable
     fn set_variable_public(&mut self, var: Variable) -> Result<(), CircuitError>;
-
-    /// Create a proof linking group in the circuit. Witness elements may be
-    /// allocated into this group in order and they will be placed into the
-    /// given layout as proof-linking gates (or assigned a default layout if
-    /// none is provided).
-    ///
-    /// These proof linking gates are effictively of the form a(x) * 0 = 0,
-    /// where a(x) is the witness element. This allows us to prove
-    /// that the a(x) polynomial of one proof equals the a(x) polynomial of
-    /// another proof over some proof-linking domain, represented by the group
-    fn create_link_group(&mut self, id: String, layout: Option<GroupLayout>) -> LinkGroup;
 
     // --- Gate Allocation --- //
